[backend] BiCG, refactor solve/minimize
